Split Shiki grammar registry modules and update tests
