+++ src/renderer/src/lib/monaco-languages/textmate-shiki-language-loaders.ts
@@ -1,0 +1,68 @@
+import type { LanguageRegistration } from '@shikijs/types'
+
+type ShikiLanguageModule = { default: LanguageRegistration | LanguageRegistration[] }
+export type ShikiLanguageLoader = () => Promise<ShikiLanguageModule>
+export type ShikiLanguageLoaderMap = Record<string, ShikiLanguageLoader | undefined>
+
+export const SHIKI_LANGUAGE_LOADERS = {
+  tsx: () => import('@shikijs/langs/tsx'),
+  jsx: () => import('@shikijs/langs/jsx'),
+  typescript: () => import('@shikijs/langs/typescript'),
+  javascript: () => import('@shikijs/langs/javascript'),
+  jsonc: () => import('@shikijs/langs/jsonc'),
+  json: () => import('@shikijs/langs/json'),
+  markdown: () => import('@shikijs/langs/markdown'),
+  mermaid: () => import('@shikijs/langs/mermaid'),
+  css: () => import('@shikijs/langs/css'),
+  scss: () => import('@shikijs/langs/scss'),
+  sass: () => import('@shikijs/langs/sass'),
+  less: () => import('@shikijs/langs/less'),
+  stylus: () => import('@shikijs/langs/stylus'),
+  postcss: () => import('@shikijs/langs/postcss'),
+  html: () => import('@shikijs/langs/html'),
+  'html-derivative': () => import('@shikijs/langs/html-derivative'),
+  xml: () => import('@shikijs/langs/xml'),
+  python: () => import('@shikijs/langs/python'),
+  rust: () => import('@shikijs/langs/rust'),
+  go: () => import('@shikijs/langs/go'),
+  java: () => import('@shikijs/langs/java'),
+  kotlin: () => import('@shikijs/langs/kotlin'),
+  c: () => import('@shikijs/langs/c'),
+  cpp: () => import('@shikijs/langs/cpp'),
+  csharp: () => import('@shikijs/langs/csharp'),
+  ruby: () => import('@shikijs/langs/ruby'),
+  php: () => import('@shikijs/langs/php'),
+  swift: () => import('@shikijs/langs/swift'),
+  shellscript: () => import('@shikijs/langs/shellscript'),
+  bat: () => import('@shikijs/langs/bat'),
+  powershell: () => import('@shikijs/langs/powershell'),
+  yaml: () => import('@shikijs/langs/yaml'),
+  ini: () => import('@shikijs/langs/ini'),
+  toml: () => import('@shikijs/langs/toml'),
+  sql: () => import('@shikijs/langs/sql'),
+  graphql: () => import('@shikijs/langs/graphql'),
+  docker: () => import('@shikijs/langs/docker'),
+  proto: () => import('@shikijs/langs/proto'),
+  lua: () => import('@shikijs/langs/lua'),
+  r: () => import('@shikijs/langs/r'),
+  scala: () => import('@shikijs/langs/scala'),
+  dart: () => import('@shikijs/langs/dart'),
+  elixir: () => import('@shikijs/langs/elixir'),
+  erlang: () => import('@shikijs/langs/erlang'),
+  haskell: () => import('@shikijs/langs/haskell'),
+  clojure: () => import('@shikijs/langs/clojure'),
+  vue: () => import('@shikijs/langs/vue'),
+  svelte: () => import('@shikijs/langs/svelte'),
+  astro: () => import('@shikijs/langs/astro'),
+  'system-verilog': () => import('@shikijs/langs/system-verilog'),
+  verilog: () => import('@shikijs/langs/verilog'),
+  nim: () => import('@shikijs/langs/nim'),
+  terraform: () => import('@shikijs/langs/terraform'),
+  csv: () => import('@shikijs/langs/csv'),
+  tsv: () => import('@shikijs/langs/tsv'),
+  make: () => import('@shikijs/langs/make'),
+  cmake: () => import('@shikijs/langs/cmake'),
+  coffee: () => import('@shikijs/langs/coffee'),
+  json5: () => import('@shikijs/langs/json5'),
+  pug: () => import('@shikijs/langs/pug')
+} satisfies ShikiLanguageLoaderMap
